refactor: 에셋 폴더 분리 및 mantine 대응 폰트 추가
